Upgraded dependencies and moved files - Changed store location - Changed component theme location - Upgraded dependencies
