+++ package.json
@@ -49,10 +49,10 @@
     "@storybook/addon-links": "^4.1.11",
     "@storybook/addon-options": "^4.1.11",
     "@storybook/html": "^4.1.13",
-    "@types/jest": "23.3.13",
+    "@types/jest": "23.3.14",
     "@types/puppeteer": "1.12.1",
     "babel-core": "^7.0.0-bridge.0",
-    "babel-jest": "^23.4.2",
+    "babel-jest": "^24.1.0",
     "babel-loader": "^8.0.4",
     "bootstrap": "^4.3.1",
     "css-loader": "^2.1.0",
@@ -60,7 +60,7 @@
     "html-loader": "^0.5.5",
     "http-proxy-middleware": "^0.19.1",
     "husky": "^1.3.1",
-    "jest": "^23.6.0",
+    "jest": "23.6.0",
     "jest-cli": "23.6.0",
     "lint-staged": "^8.1.4",
     "marked": "^0.6.0",
