--- src/components/card/card.spec.ts
@@ -1,6 +1,0 @@
-import { Card } from './card';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Card()).toBeTruthy();
-  });
-});
