--- src/components/column/column.spec.ts
@@ -1,6 +1,0 @@
-import { Column } from './column';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Column()).toBeTruthy();
-  });
-});
