--- src/components/container/container.e2e.ts
@@ -1,9 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-describe('c-container', () => {
-  it('renders', async () => {
-    const page = await newE2EPage();
-    await page.setContent('<c-container></c-container>');
-    const element = await page.find('c-container');
-    expect(element).toHaveClass('hydrated');
-  });
-});
