--- src/components/container/container.spec.ts
@@ -1,6 +1,0 @@
-import { Container } from './container';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Container()).toBeTruthy();
-  });
-});
