--- src/components/content/content.e2e.ts
@@ -1,8 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-describe('c-content', () => {
-  it('renders', async () => {
-    const page = await newE2EPage({ url: '/' });
-    const element = await page.find('c-content');
-    expect(element).toHaveClass('hydrated');
-  });
-});
