--- src/components/content/content.spec.ts
@@ -1,6 +1,0 @@
-import { Content } from './content';
-describe('app-root', () => {
-  it('builds', () => {
-    expect(new Content()).toBeTruthy();
-  });
-});
