--- src/components/field/field.spec.ts
@@ -1,6 +1,0 @@
-import { Field } from './field';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Field()).toBeTruthy();
-  });
-});
