+++ src/components/footer/footer.e2e.ts
@@ -7,8 +7,10 @@ beforeEach(async () => {
   await page.setContent('<c-footer></c-footer>');
 });
 
-const getByTestId = async dataTestId =>
+const getByTestId = async dataTestId => {
+  console.log(dataTestId)
   await page.find(`${componentName} >>> [data-test-id="${dataTestId}"]`);
+}
 
 describe('c-footer', () => {
   it('renders', async () => {
