--- src/components/header/header.e2e.ts
@@ -1,9 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-describe('c-header', () => {
-  it('renders', async () => {
-    const page = await newE2EPage();
-    await page.setContent('<c-header></c-header>');
-    const element = await page.find('c-header');
-    expect(element).toHaveClass('hydrated');
-  });
-});
