--- src/components/header/header.spec.ts
@@ -1,6 +1,0 @@
-import { Header } from './header';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Header()).toBeTruthy();
-  });
-});
