--- src/components/icon/icon.spec.ts
@@ -1,6 +1,0 @@
-import { Icon } from './icon';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Icon()).toBeTruthy();
-  });
-});
