--- src/components/list/list.spec.ts
@@ -1,6 +1,0 @@
-import { List } from './list';
-describe('app', () => {
-  it('builds', () => {
-    expect(new List()).toBeTruthy();
-  });
-});
