--- src/components/navigation/navigation.spec.ts
@@ -1,6 +1,0 @@
-import { Navigation } from './navigation';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Navigation()).toBeTruthy();
-  });
-});
