--- src/components/router-navigation/router-navigation.e2e.ts
@@ -1,9 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-describe('router-navigation', () => {
-    it('renders', async () => {
-        const page = await newE2EPage();
-        await page.setContent('<router-navigation></router-navigation>');
-        const element = await page.find('router-navigation');
-        expect(element).toHaveClass('hydrated');
-    });
-});
