--- src/components/router-navigation/router-navigation.spec.ts
@@ -1,6 +1,0 @@
-import { RouterNavigation } from './router-navigation';
-describe('app', () => {
-    it('builds', () => {
-        expect(new RouterNavigation()).toBeTruthy();
-    });
-});
