--- src/components/row/row.e2e.ts
@@ -1,9 +1,0 @@
-import { newE2EPage } from '@stencil/core/testing';
-describe('c-row', () => {
-  it('renders', async () => {
-    const page = await newE2EPage();
-    await page.setContent('<c-row></c-row>');
-    const element = await page.find('c-row');
-    expect(element).toHaveClass('hydrated');
-  });
-});
