--- src/components/row/row.spec.ts
@@ -1,6 +1,0 @@
-import { Row } from './row';
-describe('app', () => {
-  it('builds', () => {
-    expect(new Row()).toBeTruthy();
-  });
-});
