--- src/components/theme/theme.spec.ts
@@ -1,6 +1,0 @@
-import { Theme } from './theme';
-describe('app-root', () => {
-  it('builds', () => {
-    expect(new Theme()).toBeTruthy();
-  });
-});
