--- src/components/user-repos/user-repos.spec.ts
@@ -1,6 +1,0 @@
-import { UserRepos } from './user-repos';
-describe('app', () => {
-  it('builds', () => {
-    expect(new UserRepos()).toBeTruthy();
-  });
-});
