+++ src/global.ts
